Adding README and refactoring code.

+++ rss_reader/to_pdf_converter.py
@@ -1,3 +1,4 @@
+"""Module which provides interface of translating news to .pdf."""
 import os
 from fpdf import FPDF
 from image_handler import save_image_by_url
@@ -6,32 +7,41 @@ from image_handler import save_image_by_url
 TITLE_IMG_NAME = 'title.png'
 
 
-class PDF(FPDF):
-	# def __init__(self, title: str, title_img_url: str):
+def _convert_to_latin_1(string: str) -> str:
+	"""Convert to latin-1 encoding."""
+	return string.encode('latin-1', 'replace').decode('latin-1')
 
+
+class PDF(FPDF):
+	"""Class, which allows to translate news to .pdf format."""
 
 	def set_meta_info(self, title: str, title_img_url: str):
-		self.title = title
+		"""Set information of rss-resource."""
+		self.title = _convert_to_latin_1(title)
 		self.title_img_url = title_img_url
 
 		self.iter = 0
 
 
-	def garbage_collector(self):
-		'''Removing temp files'''
+	def _garbage_collect(self):
+		"""Remove temp img-files."""
 		for index in range(self.iter):
 			os.remove('temp' + str(index) + '.png')
-		os.remove(TITLE_IMG_NAME)
+		if self.title_img_url != '':
+			os.remove(TITLE_IMG_NAME)
 
 
 	def write_to_file(self, filepath: str):
-		self.garbage_collector()
-		self.output(u'news.pdf')
+		"""Write to file pdf items."""
+		self._garbage_collect()
+		self.output('news.pdf')
 
 
 	def header(self):
-		save_image_by_url(self.title_img_url, TITLE_IMG_NAME)
-		self.image(TITLE_IMG_NAME, 10, 8, 33)
+		"""Set header of each page."""
+		if self.title_img_url != '':
+			save_image_by_url(self.title_img_url, TITLE_IMG_NAME)
+			self.image(TITLE_IMG_NAME, 10, 8, 33)
 
 		self.set_font("Arial", 'B', size=12)
 		self.cell(100)
@@ -41,6 +51,7 @@ class PDF(FPDF):
 
 
 	def footer(self):
+		"""Set footer of each page."""
 		self.set_y(-10)
  
 		self.set_font('Arial', 'I', 8)
@@ -50,33 +61,39 @@ class PDF(FPDF):
 
 
 	def _add_title_of_news(self, title: str) -> None:
+		"""Insert title of piece of news."""
 		self.set_font('Times', 'B', size=13)
-		self.multi_cell(0,10, txt=title)
+		self.multi_cell(0,10, txt=_convert_to_latin_1(title))
 
 
 	def _add_date_of_news(self, date: str) -> None:
+		"""Insert date of piece of news."""
 		self.set_font('Times', size=12)
-		self.multi_cell(0,10, txt=date)
+		self.multi_cell(0,10, txt=_convert_to_latin_1(date))
 
 
 	def _add_link_of_news(self, link: str) -> None:
+		"""Insert link of piece of news."""
 		self.set_font('Arial','I', size=11)
-		self.multi_cell(0,10, txt=link)
+		self.multi_cell(0,10, txt=_convert_to_latin_1(link))
 
 
 	def _add_content_of_news(self, content: str) -> None:
+		"""Insert content of piece of news."""
 		self.set_font('Times', size=12)
-		self.multi_cell(0,10, txt=content)
+		self.multi_cell(0,10, txt=_convert_to_latin_1(content))
 
 
 	def _add_img_of_news(self, img_url: str) -> None:
+		"""Insert image of piece of news."""
 		save_image_by_url(img_url, 'temp' + str(self.iter) + '.png')
-		self.image('temp' + str(self.iter) + '.png')
+		self.image('temp' + str(self.iter) + '.png', None, None)
 
 		self.iter += 1
 
 
 	def add_piece_of_news(self, title: str, date:str, link: str, img_url: str, content: str):
+		"""Insert piece of news to pdf."""
 		self.alias_nb_pages()
 		self.add_page()
